+++ frontend/pages/_app.tsx
@@ -1,8 +1,14 @@
 import type { AppProps } from 'next/app'
+import { Layout } from '../components/layout'
 import '../styles/globals.css'
+import { AuthProvider } from '../lib/AuthContext'
 
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp 
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <AuthProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </AuthProvider>
+  )
+}